+++ gears/environment.py
@@ -405,8 +405,8 @@ class Environment(object):
         """Save handled public assets to :attr:`root` directory."""
         for asset_attributes, absolute_path in self.list('.', recursive=True):
             logical_path = os.path.normpath(asset_attributes.logical_path)
-            if self.is_public(logical_path):
-                asset = build_asset(self, logical_path)
+            asset = build_asset(self, logical_path)
+            if asset.is_public:
                 source = bytes(asset)
                 self.save_file(logical_path, source, asset.gzippable)
                 if self.fingerprinting: